Move About, App, Play and Settings components to live within routes folder
