+++ src/routes/About/About.js
@@ -1,5 +1,4 @@
-import React from "react";
-import Button from "../Button/Button";
+import Button from "../../components/Button/Button";
 
 
 function About() {
