+++ src/routes/Play/Play.test.tsx
@@ -1,0 +1,23 @@
+// import {render, screen} from '@testing-library/react';
+// import userEvent from '@testing-library/user-event';
+// import Home from './Home.js';
+
+
+// beforeEach(() => render(<Home />));
+
+// test('render apps title name', () => {
+//   const title = screen.getByRole('heading');
+//   expect(title).toHaveTextContent('Memory Match');
+// });
+
+// test('renders the welcome message', () => {
+//   const welcomeMsg = screen.getByTestId('welcomeMsg');
+//   expect(welcomeMsg).toHaveTextContent('Are you ready to test your memory?');
+// });
+
+// test('renders a button with text Play', () => {
+//   const button = screen.getByRole('button');
+//   expect(button).toHaveTextContent('Play');
+// });
+
+export {}
